python/requirements.txt
fastapi
uvicorn
requests
httpx
python-dotenv
